// src/types/@gearbox-protocol/router/contracts/interfaces/IClosePathResolver.ts
export type MultiCallStruct = { target: string; callData: BytesLike };